Перераспределил файлы по пакетам. Также создал папку utils, куда сделал доступ для получения SnackBar'a и функции конвертации полученных данных для AppState.
