+++ app/src/main/java/com/example/kotlin_lesson_1/viewModel/MainViewModel.kt
@@ -1,10 +1,9 @@
 package com.example.kotlin_lesson_1.viewModel
 
-import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.example.kotlin_lesson_1.model.Repository
-import com.example.kotlin_lesson_1.model.RepositoryImpl
+import com.example.kotlin_lesson_1.repository.Repository
+import com.example.kotlin_lesson_1.repository.RepositoryImpl
 import java.lang.Thread.sleep
 
 class MainViewModel(
